Update rules_jvm_external to use the Starlark version of aar_import automatically if the version of Bazel used does not contain the native version of aar_import.
This also updates the version of rules_android to the latest commit, because
version 0.1.1 is basically just wrappers around the native versions of the rules.

This also disbles `--experimental_sibling_repository_layout` added in
https://github.com/bazelbuild/rules_jvm_external/commit/b6631f90a46c23be32000bf14b8e817f44017cc4
because the latest rules_android uses protos in go, and there is a problem compiling protos in go
with `--experimental_sibling_repository_layout`: https://github.com/bazelbuild/rules_go/issues/3947

Fixes #1139.